用户模块基本功能已完成，登录控制待完成

+++ src/main/java/com/yjy/model/Teacher.java
@@ -48,9 +48,9 @@ public class Teacher {
     /**
      * 创建时间
      */
-    private BigInteger teacherRegtime;
+    private Long teacherRegtime;
     /**
      * 修改时间
      */
-    private BigInteger updateTime;
+    private Long updateTime;
 }
